+++ client/src/ui-components/ui/form.tsx
@@ -1,10 +1,12 @@
 import * as React from 'react'
-import { Controller, ControllerProps, FieldPath, FieldValues, FormProvider, useFormContext } from 'react-hook-form'
+import { Controller, ControllerProps, FieldPath, FieldValues, FormProvider } from 'react-hook-form'
 
 import * as LabelPrimitive from '@radix-ui/react-label'
 import { Slot } from '@radix-ui/react-slot'
 import { Label } from '@ui-components/label'
 import { cn } from '@utils/tailwind'
+
+import { useFormField } from './use-form-field'
 
 const Form = FormProvider
 
@@ -15,7 +17,7 @@ type FormFieldContextValue<
     name: TName
 }
 
-const FormFieldContext = React.createContext<FormFieldContextValue>({} as FormFieldContextValue)
+export const FormFieldContext = React.createContext<FormFieldContextValue>({} as FormFieldContextValue)
 
 const FormField = <
     TFieldValues extends FieldValues = FieldValues,
@@ -30,34 +32,13 @@ const FormField = <
     )
 }
 
-const useFormField = () => {
-    const fieldContext = React.useContext(FormFieldContext)
-    const itemContext = React.useContext(FormItemContext)
-    const { getFieldState, formState } = useFormContext()
 
-    const fieldState = getFieldState(fieldContext.name, formState)
-
-    if (!fieldContext) {
-        throw new Error('useFormField should be used within <FormField>')
-    }
-
-    const { id } = itemContext
-
-    return {
-        id,
-        name: fieldContext.name,
-        formItemId: `${id}-form-item`,
-        formDescriptionId: `${id}-form-item-description`,
-        formMessageId: `${id}-form-item-message`,
-        ...fieldState,
-    }
-}
 
 type FormItemContextValue = {
     id: string
 }
 
-const FormItemContext = React.createContext<FormItemContextValue>({} as FormItemContextValue)
+export const FormItemContext = React.createContext<FormItemContextValue>({} as FormItemContextValue)
 
 const FormItem = React.forwardRef<HTMLDivElement, React.HTMLAttributes<HTMLDivElement>>(
     ({ className, ...props }, ref) => {
@@ -133,4 +114,4 @@ const FormMessage = React.forwardRef<HTMLParagraphElement, React.HTMLAttributes<
 )
 FormMessage.displayName = 'FormMessage'
 
-export { useFormField, Form, FormItem, FormLabel, FormControl, FormDescription, FormMessage, FormField }
+export { Form, FormItem, FormLabel, FormControl, FormDescription, FormMessage, FormField }
