+++ server/package.json
@@ -24,6 +24,7 @@
         "@google-cloud/pubsub": "^3.7.1",
         "@hubspot/api-client": "^9.0.0",
         "@sendgrid/mail": "^7.7.0",
+        "@sparticuz/chromium": "^121.0.0",
         "@trpc/server": "^10.43.3",
         "acuityscheduling": "^0.1.9",
         "firebase-admin": "^11.11.0",
