+++ server/src/party-bookings/invitations/freckles.html
@@ -9,10 +9,16 @@
 
 
     <style>
-        body {
+        html {
             width: 210mm;
             height: 290mm;
             margin: 0;
+        }
+
+        body {
+            width: 210mm;
+            height: 290mm;
+            margin: 0px;
         }
