+++ server/src/trpc/trpc.adapter.ts
@@ -1,4 +1,4 @@
-import { logger } from 'firebase-functions/v2'
+import { MemoryOption, logger } from 'firebase-functions/v2'
 import { onRequest } from 'firebase-functions/v2/https'
 
 import { AnyRouter } from '@trpc/server'
@@ -6,9 +6,9 @@ import { createHTTPHandler } from '@trpc/server/adapters/standalone'
 
 import { createContext } from './trpc'
 
-export function onRequestTrpc<TRouter extends AnyRouter>(router: TRouter) {
+export function onRequestTrpc<TRouter extends AnyRouter>(router: TRouter, memory?: MemoryOption) {
     return onRequest(
-        { region: 'australia-southeast1', cors: true },
+        { region: 'australia-southeast1', cors: true, memory: memory || '256MiB' },
         createHTTPHandler({
             router,
             createContext,
